requirements.txt
torch
transformers
faiss-cpu
pandas
numpy
pillow
tqdm
gradio
pyarrow
sentencepiece
